Move authentication and authorization failure tests to another test file
